@@ -12,12 +12,12 @@
     "@types/react": "^18.0.5",
     "@types/react-dom": "^18.0.0",
     "@types/react-router-dom": "^5.3.3",
-    "axios": "^0.26.1",
+    "axios": "^0.27.2",
     "materialize-css": "^1.0.0-rc.2",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-redux": "^7.2.8",
-    "react-router-dom": "^5.1.2",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "^5.0.1",
     "typescript": "^4.6.3"
   },
