@@ -1,15 +1,16 @@
 import React, { MouseEvent, useContext } from "react";
-import { NavLink, useHistory } from "react-router-dom";
+import { NavLink, useNavigate, useLocation } from "react-router-dom";
 import { AuthContext } from "../context/AuthContext";
 
 export const Navbar = () => {
-  const history = useHistory();
+  const navigate = useNavigate();
   const auth = useContext(AuthContext);
+  const { pathname } = useLocation();
 
   const logoutHandler = (event: MouseEvent<HTMLElement>) => {
     event.preventDefault();
     auth.logout();
-    history.push("/");
+    navigate("/");
   };
 
   return (
@@ -17,19 +18,48 @@ export const Navbar = () => {
       <div className="nav-wrapper blue darken-1" style={{ padding: "0 2rem" }}>
         <span className="brand-logo">Сокращение ссылок</span>
         <ul id="nav-mobile" className="right hide-on-med-and-down">
-          <li>
+          <li
+            style={{
+              backgroundColor: pathname === "/create" ? "blue" : "#1E88E5",
+            }}
+          >
             <NavLink to="/create">Создать</NavLink>
           </li>
-          <li>
+          <li
+            style={{
+              backgroundColor: pathname === "/file" ? "blue" : "#1E88E5",
+            }}
+          >
+            <NavLink to="/file">Загрузить файл</NavLink>
+          </li>
+          <li
+            style={{
+              backgroundColor: pathname === "/links/fetch" ? "blue" : "#1E88E5",
+            }}
+          >
             <NavLink to="/links/fetch">Fetch</NavLink>
           </li>
-          <li>
+          <li
+            style={{
+              backgroundColor:
+                pathname === "/links/create-async-thunk" ? "blue" : "#1E88E5",
+            }}
+          >
             <NavLink to="/links/create-async-thunk">Create async thunk</NavLink>
           </li>
-          <li>
+          <li
+            style={{
+              backgroundColor: pathname === "/links/RTK" ? "blue" : "#1E88E5",
+            }}
+          >
             <NavLink to="/links/RTK">RTK</NavLink>
           </li>
-          <li>
+          <li
+            style={{
+              backgroundColor:
+                pathname === "/links/json-server" ? "blue" : "#1E88E5",
+            }}
+          >
             <NavLink to="/links/json-server">json-server</NavLink>
           </li>
           <li>
