@@ -1,12 +1,20 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot, hydrateRoot } from "react-dom/client";
 import "./index.css";
-import App from "./App";
+import { MountDomEffect } from "./MountDomEffect";
 import * as serviceWorker from "./serviceWorker";
 
-ReactDOM.render(<App />, document.getElementById("root"));
+const container = document.getElementById("root");
+const root = createRoot(container!); // createRoot(container!) if you use TypeScript
+
+root.render(<MountDomEffect />);
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
 // Learn more about service workers: https://bit.ly/CRA-PWA
 serviceWorker.unregister();
+
+// For study of hydrate
+// import { hydrateRoot } from 'react-dom/client';
+// const container = document.getElementById('root');
+// const root = hydrateRoot(container, <MountDomEffect />);
