@@ -1,9 +1,6 @@
 const { Router } = require("express");
-const bcrypt = require("bcryptjs");
-const config = require("config");
-const jwt = require("jsonwebtoken");
-const { check, validationResult } = require("express-validator");
-const User = require("../models/User.js");
+const { check } = require("express-validator");
+const AuthController = require("../Controllers/AuthController.js");
 const router = Router();
 
 router.post(
@@ -14,39 +11,7 @@ router.post(
       min: 6,
     }),
   ],
-  async (req, res) => {
-    try {
-      const errors = validationResult(req);
-
-      if (!errors.isEmpty()) {
-        return res.status(400).json({
-          errors: errors.array(),
-          message: "Некорректный данные при регистрации",
-        });
-      }
-
-      const { email, password } = req.body;
-
-      const candidate = await User.findOne({ email });
-
-      if (candidate) {
-        return res
-          .status(400)
-          .json({ message: "Такой пользователь уже существует" });
-      }
-
-      const hashedPassword = await bcrypt.hash(password, 12);
-      const user = new User({ email, password: hashedPassword });
-
-      await user.save();
-
-      res.status(201).json({ message: "Пользователь создан" });
-    } catch (e) {
-      res
-        .status(500)
-        .json({ message: "Что-то пошло не так, попробуйте снова" });
-    }
-  }
+  AuthController.register
 );
 
 router.post(
@@ -55,48 +20,7 @@ router.post(
     check("email", "Введите корректный email").normalizeEmail().isEmail(),
     check("password", "Введите пароль").exists(),
   ],
-  async (req, res) => {
-    try {
-      const errors = validationResult(req);
-
-      if (!errors.isEmpty()) {
-        return res.status(400).json({
-          errors: errors.array(),
-          message: "Некорректный данные при входе в систему",
-        });
-      }
-
-      const { email, password } = req.body;
-
-      const user = await User.findOne({ email });
-
-      if (!user) {
-        return res.status(400).json({ message: "Пользователь не найден" });
-      }
-
-      const isMatch = await bcrypt.compare(password, user.password);
-
-      if (!isMatch) {
-        return res
-          .status(400)
-          .json({ message: "Неверный пароль, попробуйте снова" });
-      }
-
-      const token = jwt.sign(
-        { userId: user.id },
-        config.get("jwtAccessSecret")
-        // {
-        //   expiresIn: "1h",
-        // }
-      );
-
-      res.json({ token, userId: user.id });
-    } catch (e) {
-      res
-        .status(500)
-        .json({ message: "Что-то пошло не так, попробуйте снова" });
-    }
-  }
+  AuthController.login
 );
 
 module.exports = router;
